requirements.txt
agate==1.9.1
aiosqlite==0.22.1
alembic==1.18.4
amplitude-analytics==1.2.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
apprise==1.9.7
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgi-lifespan==2.1.0
asttokens==3.0.1
asyncpg==0.31.0
attrs==25.4.0
babel==2.18.0
beartype==0.22.9
beautifulsoup4==4.14.3
cachetools==7.0.1
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
colorama==0.4.6
comm==0.2.3
coolname==3.0.0
croniter==6.0.0
cryptography==46.0.5
daff==1.4.2
dateparser==1.3.0
dbt-adapters==1.22.6
dbt-common==1.37.2
dbt-core==1.11.6
dbt-duckdb==1.10.1
dbt-extractor==0.6.0
dbt-protos==1.0.431
dbt-semantic-interfaces==0.9.0
debugpy==1.8.20
decorator==5.2.1
deepdiff==8.6.1
docker==7.1.0
duckdb==1.4.4
exceptiongroup==1.3.1
executing==2.2.1
fakeredis==2.34.0
fastapi==0.129.0
fastexcel==0.19.0
fsspec==2026.2.0
graphviz==0.21
greenlet==3.3.1
griffe==2.0.0
griffecli==2.0.0
griffelib==2.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.15.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
isodate==0.7.2
jedi==0.19.2
Jinja2==3.1.6
jinja2-humanize-extension==0.4.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter_client==8.8.0
jupyter_core==5.9.1
leather==0.4.1
lupa==2.6
Mako==1.3.10
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mashumaro==3.14
matplotlib-inline==0.2.1
mdurl==0.1.2
minio==7.2.20
more-itertools==10.8.0
msgpack==1.1.2
nest-asyncio==1.6.0
networkx==3.6.1
oauthlib==3.3.1
opentelemetry-api==1.39.1
orderly-set==5.5.0
orjson==3.11.7
packaging==25.0
parsedatetime==2.6
parso==0.8.6
pathspec==0.12.1
platformdirs==4.9.2
pluggy==1.6.0
polars==1.38.1
polars-runtime-32==1.38.1
prefect==3.6.18
prometheus_client==0.24.1
prompt_toolkit==3.0.52
protobuf==6.33.5
psutil==7.2.2
pure_eval==0.2.3
py-key-value-aio==0.4.4
pycparser==3.0
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-extra-types==2.11.0
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydocket==0.17.8
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.2
pywin32==311
PyYAML==6.0.3
pyzmq==27.1.0
readchar==4.2.1
redis==7.2.0
referencing==0.37.0
regex==2026.2.19
requests==2.32.5
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==14.3.3
rpds-py==0.30.0
ruamel.yaml==0.19.1
ruamel.yaml.clib==0.2.15
semver==3.0.4
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
snowplow-tracker==1.1.0
sortedcontainers==2.4.0
soupsieve==2.8.3
SQLAlchemy==2.0.46
sqlparse==0.5.4
stack-data==0.6.3
starlette==0.52.1
text-unidecode==1.3
toml==0.10.2
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
typer==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.41.0
wcwidth==0.6.0
websockets==16.0
whenever==0.9.5
zipp==3.23.0
